libgis: Enable the C99 bool type (#1567)
Enables globally the C99 'bool' type with its values 'true' and 'false'.

The old boolean constants 'TRUE' and 'FALSE' are still valid but
redefined to 'true' and 'false' to emphasize their equivalence.

(Re-)definitions of TRUE/FALSE in modules are removed.

This change was originally triggered by a conflict of v.hull -- defining
a local bool type -- with GDAL 3.3 which introduced the inclusion of
stdbool.h (#1563).