libpython: Save and load benchmark results (#1711)
Series of functions to save benchmark results to JSON, load them back, and combine them together.

This needs some command line interface to make it convenient, but I used it in a custom benchmark script.

The JSON save ensures that the saved results are a dictionary not just a list which makes the storage format more extensible as first level can contain additional keys.