ci: Remove duplication in Ubuntu workflow, keep test job only (#1749)
Remove duplication introduced with PDAL: matrix.os has the name of the the system which is used in the YAML.
This basically returns the CI YAML files state to what it was before the change, but
keeping the addition of a new build script and version-specific naming for the scripts.

Remove the separate Ubuntu build, keep only the test job, rename CI to Ubuntu.
There is no need for a separate build anymore since GCC workflows now do that.

Additionally, fix CentOS label, add empty lines between list items (maybe helpful; allowed by prettier).