raster: Work with any mask name (r.surf.contour, r.random.cells, r.random.surface) (#4634)
Use Rast_mask_status instead of hardcoded name MASK for r.surf.contour, r.random.cells, and r.random.surface. Raster mask if present, is loaded in a special way in these tools. After this change, the name of the mask is retrieved from the library (that's the important part) and the library is at the same time used to test the presence of the mask (as opposed to testing presence of raster directly, but that's just more convenient rather than conceptual).

The change in r.random.surface code switching the if and else branches is to make it easier to write and more consistent with the other two.